chore: library cleanup — clean imports, feature names, genuine CFs, dead code
Tier 1:
- import microfactual no longer emits DeprecationWarnings: functional impls
  relocated to core.processing / models.training / visualization.roc_io; the
  deprecated data_processing/modeling/visualisation modules are thin re-export
  shims that warn only on direct import
- explain_counterfactual drops DiCE rows that don't actually flip the prediction
  (no more zero-change 'counterfactuals' / validity < 100%)

T2.3 (feature names):
- MicrobiomeClassifier.predict/predict_proba stop stripping DataFrame column
  names -> removes sklearn 'X has no valid feature names' UserWarning and keeps
  downstream sklearn tooling working (test warnings 32 -> 2)

T3.2 (deprecation shims): consolidated with clear v1.0 removal timeline.

Tier 3:
- removed unused BaseModel ABC (never implemented) and orphaned src/microbiome_ml

2 new regression tests; 82 tests pass; ruff clean; docs build.
