+++ src/microfactual/core/__init__.py
@@ -1,6 +1,5 @@
-"""Core data structures and base classes for microfactual."""
+"""Core data structures for microfactual."""
 
-from .base import BaseModel
 from .dataset import MicrobiomeDataset
 
-__all__ = ["MicrobiomeDataset", "BaseModel"]
+__all__ = ["MicrobiomeDataset"]
